V2022.06.04 - Live Preview
Made an update last year and forgot to upload it. And since I gave Allen permission to use it on the CC discord server for his injector site revamp, figured I'd update it to the last change this for prosperity. 

Basically the main change was making it so that it would automatically update the preview when the file changed. More accurately it was periodically soft reloading the image sheet.

If I were to make a new update I would implement a file selection instead of requiring the user to have it be ./sprite_sheet.png in the same directory as the HTML file.